requirements.txt
streamlit
streamlit-option-menu
pandas
numpy
plotly
yfinance
pillow
pypfopt
empyrical
scipy
python-dateutil


